+++ src/xyz/oihana/schema/constants/traits/business/documents/ProofOfDeliveryTrait.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\business\documents;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\business\documents\ProofOfDelivery} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\business\documents
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.3.0
+ */
+trait ProofOfDeliveryTrait
+{
+    const string DATE             = 'date' ;
+    const string DISCREPANCY_NOTE = 'discrepancyNote' ;
+    const string SIGNATORY        = 'signatory' ;
+}